flask-server/requirements.txt
flask==2.3.3
flask-cors==4.0.0
selenium==4.15.0
webdriver-manager==4.0.1
weasyprint==60.0
requests==2.31.0
python-dotenv==1.0.0